Add unit tests for question-player component (#13322)
* Changed directive to component

* Save file

* Added some tests

* 100% coverage

* Lint err

* Fix e2e
