+++ core/templates/components/question-directives/question-player/question-player.component.html
