+++ core/templates/pages/review-test-page/review-test-page.component.ts
@@ -24,7 +24,7 @@ require(
   'background-banner.component.ts');
 require(
   'components/question-directives/question-player/' +
-  'question-player.directive.ts');
+  'question-player.component.ts');
 require('interactions/interactionsQuestionsRequires.ts');
 require('objects/objectComponentsRequiresForPlayers.ts');
 require('pages/interaction-specs.constants.ajs.ts');
